+++ src/trader/apis/bva.ts
@@ -2,8 +2,9 @@ import axios from "axios"
 
 import env from "../env"
 import logger from "../../logger"
-import { BvaCommand, TradeOpen } from "../types/bva"
+import { BvaCommand, TradeOpen, TradeOpenJson } from "../types/bva"
 
+// Gets a list of all open trades for the user
 export async function getTradeOpenList(): Promise<TradeOpen[]> {
     return new Promise((resolve, reject) => {
         axios
@@ -32,6 +33,38 @@ export async function getTradeOpenList(): Promise<TradeOpen[]> {
             })
             .catch((reason) => {
                 logger.error(`Failed to get open trades: ${reason}`)
+                reject(reason)
+            })
+    })
+}
+
+// Gets a list of all open trades for the source strategy
+export async function getStratTradeOpenList(stratid: string): Promise<TradeOpen[]> {
+    return new Promise((resolve, reject) => {
+        axios
+            .get(
+                `https://api-bva.herokuapp.com/api/opensignals?id=${stratid}`
+            )
+            .then((response) => {
+                logger.silly(`Fetched strategy trades: ${JSON.stringify(response.data)}`)
+                const data: TradeOpenJson[] = response.data
+                const tradeOpens = data.map(
+                    (tradeOpenJson) => new TradeOpen(tradeOpenJson)
+                )
+
+                const tradeOpenSymbols = tradeOpens
+                    .map((tradeOpen) => tradeOpen.symbol)
+                    .join(", ")
+                logger.debug(
+                    `Fetched ${tradeOpens.length} open trades for ${stratid}${
+                        tradeOpenSymbols && ": " + tradeOpenSymbols
+                    }.`
+                )
+
+                resolve(tradeOpens)
+            })
+            .catch((reason) => {
+                logger.error(`Failed to get strategy open trades: ${reason}`)
                 reject(reason)
             })
     })